fix endpoints according to api rest good practices
