+++ backend/src/main/java/com/mentora/backend/controller/ForumController.java
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import com.mentora.backend.responses.DtApiResponse;
 import com.mentora.backend.responses.GetForumResponse;
-import com.mentora.backend.responses.GetPostResponse;
 
 @RestController
 @RequestMapping("/forum")
@@ -81,38 +80,6 @@ public class ForumController {
                 HttpStatus.OK.value(),
                 "Lista de posts obtenida correctamente",
                 forumResponse
-            ));
-        } catch (ResponseStatusException e) {
-            return ResponseEntity.status(e.getStatusCode()).body(new DtApiResponse<>(
-                false,
-                e.getStatusCode().value(),
-                e.getReason(),
-                null
-            ));
-        }
-    }
-
-    @Operation(
-            summary = "Obtener un post de un foro",
-            description = "Devuelve el post y datos del foro.",
-            security = @SecurityRequirement(name = "bearerAuth"))
-    @ApiResponse(responseCode = "200", description = "Post obtenido correctamente")
-    @ApiResponse(responseCode = "404", description = "Foro o post no encontrado")
-    @ApiResponse(responseCode = "500", description = "Error interno del servidor")
-    @PreAuthorize("hasAnyRole('PROFESOR','ESTUDIANTE')")
-    @GetMapping("/{forumId}/post/{postId}")
-    public ResponseEntity<DtApiResponse<GetPostResponse>> getPost(
-            @PathVariable Long forumId,
-            @PathVariable Long postId
-    ) {
-        try {
-            GetPostResponse postResponse = forumService.getPost(forumId, postId);
-
-            return ResponseEntity.ok(new DtApiResponse<>(
-                true,
-                HttpStatus.OK.value(),
-                "Post obtenido correctamente",
-                postResponse
             ));
         } catch (ResponseStatusException e) {
             return ResponseEntity.status(e.getStatusCode()).body(new DtApiResponse<>(
