+++ backend/src/main/java/com/mentora/backend/requests/CreatePostResponseRequest.java
@@ -1,0 +1,8 @@
+package com.mentora.backend.requests;
+
+public class CreatePostResponseRequest {
+  private String message;
+
+  public String getMessage() { return message; }
+  public void setMessage(String message) { this.message = message; }
+}
