+++ backend/src/main/java/com/mentora/backend/service/ForumService.java
@@ -6,7 +6,6 @@ import com.mentora.backend.dt.DtForum;
 import com.mentora.backend.model.*;
 import com.mentora.backend.repository.*;
 import com.mentora.backend.responses.GetForumResponse;
-import com.mentora.backend.responses.GetPostResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -100,20 +99,6 @@ public class ForumService {
 
         List<Post> posts = postRepository.findByForum_IdOrderByCreatedDateDesc(forumId);
         return new GetForumResponse(getDtForum(forum), posts.stream().map(this::getDtPost).toList());
-    }
-
-    public GetPostResponse getPost(Long forumId, Long postId) {
-        Forum forum = forumRepository.findById(forumId)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Foro no encontrado"));
-
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Post no encontrado"));
-
-        if (!post.getForum().getId().equals(forum.getId())) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "El post no pertenece al foro");
-        }
-
-        return new GetPostResponse(getDtForum(forum), getDtPost(post));
     }
 
     private DtPost getDtPost(Post post) {
